added sequel injection protection. added sequelize.escape to all sequelize statements that use input values

+++ server/controllers/userFortunes.js
@@ -31,9 +31,9 @@ module.exports = {
       try {
          const { text, userAdded, userId } = req.body
          await Fortune.create({
-            text,
-            userAdded,
-            userId,
+            text: sequelize.escape(text),
+            userAdded: sequelize.escape(userAdded),
+            userId: sequelize.escape(userId),
          })
          res.status(200).send('Added fortune succesfully')
       } catch (err) {
@@ -47,7 +47,7 @@ module.exports = {
       try {
          const { sub } = await verifyAccessToken(token)
          if (sub) {
-            await User.update({ username }, { where: { id: sub } })
+            await User.update({ username }, { where: { id: sequelize.escape(sub) } })
             res.status(200).send('username updated successfully')
          } else return res.status(200).send('not authenticated')
       } catch (err) {
@@ -59,14 +59,14 @@ module.exports = {
    toggleFavorite: async (req, res) => {
       const { userId, fortuneId } = req.body
       try {
-         let found = await Favorite.findOne({ where: { userId, fortuneId } })
+         let found = await Favorite.findOne({ where: { userId: sequelize.escape(userId), fortuneId: sequelize.escape(fortuneId) } })
          console.log(found)
          if (found) {
-            let dest = await Favorite.destroy({ where: { userId, fortuneId } })
+            let dest = await Favorite.destroy({ where: { userId: sequelize.escape(userId), fortuneId: sequelize.escape(fortuneId) } })
             console.log(dest)
             return res.status(200).send('favorite removed')
          } else {
-            let created = await Favorite.create({ userId, fortuneId })
+            let created = await Favorite.create({ userId: sequelize.escape(userId), fortuneId: sequelize.escape(fortuneId) })
             console.log('CREATED', created)
             return res.status(200).send('favorite created')
          }
@@ -79,7 +79,7 @@ module.exports = {
    determineFav: async (req, res) => {
       const { userId, fortuneId } = req.body
       try {
-         let found = await Favorite.findOne({ where: { userId, fortuneId } })
+         let found = await Favorite.findOne({ where: { userId: sequelize.escape(userId), fortuneId: sequelize.escape(fortuneId) } })
          if (found) return res.status(200).send(true)
          else return res.status(200).send(false)
       } catch (err) {
@@ -103,7 +103,7 @@ module.exports = {
       const { id } = req.body
       try {
          let favs = await Favorite.findAll({
-            where: { userId: id },
+            where: { userId: sequelize.escape(id) },
             include: [{model: Fortune, include: [{model: User}]}]
          })
          res.status(200).send(favs)
@@ -117,7 +117,7 @@ module.exports = {
       const { id } = req.params
       console.log('id', id)
       try {
-         let deleted = await Favorite.destroy({where: {id}})
+         let deleted = await Favorite.destroy({where: { id: sequelize.escape(id) }})
          console.log(deleted)
          
       } catch (err) {
